[consensus] breaking up safety rules crate
Preparing for an interface for SafetyRules by separating these data
structures into their own files for logical separation.
